Initial Commit - Thesis
- Created a permutation builder to calculate the
overall cost of pairing players together in a 1v1
scenario. This naive solution will not work for a NvN
scenario since the time complexity is exponential and
quickly gets to be far too long for a proper match
making system. This problem is NP complete thus a greedy
solution would need to be implemented or another way of
reducing the time complexity would need to be introduced.