Added Data from CSV to Implementation
- Adding the CSV data shows the NP-completeness of
the problem. With this implementation, it gets a recursion
error for going too deep (this is because the data is about
3400 entries). When reducing the number of entries just down
to 10, it ends up taking a large amount of time to calculate.
From 8 players -> 10 players, the execution time difference is
gigantic. This is because is now needs to show every permutation
for 5 games rather than 4. Obviously by doing using a larger set
of players like the dataset, it is not possible to find the
optimal match in a reasonable amount of time.

- Note: The current implementation is far from optimized and does
a lot of calculations that are redundant however optimizing this
solution would still result in an unreasonable execution time.